chore. 형 변환 타입 명시

+++ src/main/java/model/LottoResultAnalyzer.java
@@ -36,7 +36,7 @@ public class LottoResultAnalyzer {
 
     public double calculateProfitRate(PurchaseAmount purchaseAmount) {
         long totalWinning = result.entrySet().stream()
-            .mapToLong(entry -> entry.getKey().getPrize() * entry.getValue())
+            .mapToLong(entry -> (long) entry.getKey().getPrize() * entry.getValue())
             .sum();
 
         return (double) totalWinning / purchaseAmount.getAmount();
